intent-search-project/requirements.txt
transformers==4.31.0
sentence-transformers==2.2.2
pinecone-client==3.2.0
pandas==2.0.3
datasets==2.14.0
torch==2.0.1
python-dotenv==1.0.0
huggingface_hub==0.16.4
seqeval==1.2.2
apache-airflow==2.7.2
mlflow==2.6.0
psycopg2-binary==2.9.6
uvicorn==0.23.2
fastapi==0.103.0
pendulum==2.1.2
email-validator